app/requirements.txt
django-allauth==0.22.0
django-suit==0.2.13
django-tastypie==0.12.2
django==1.8.3
factory-boy==2.5.2
psycopg2==2.6.1
pytest-cov==1.8.1
pytest-django==2.8.0
pytest-flakes==1.0.0
pytest-pep8==1.0.6
pytest-xdist==1.12
pytest==2.7.2
